Create SupportTicketFactory
